Dx8 updated from 5.7.10 to 6.1.2.

+++ docroot/modules/contrib/dx8/modules/cohesion_website_settings/src/Entity/FontLibrary.php
@@ -2,11 +2,8 @@
 
 namespace Drupal\cohesion_website_settings\Entity;
 
-use Drupal\cohesion\Entity\CohesionConfigEntityBase;
 use Drupal\Core\Entity\EntityStorageInterface;
-use Drupal\Component\Serialization\Json;
 use Drupal\Core\Cache\Cache;
-use Drupal\cohesion_website_settings\Plugin\Api\WebsiteSettingsApi;
 use Drupal\cohesion\Entity\CohesionSettingsInterface;
 
 /**
@@ -38,6 +35,17 @@ use Drupal\cohesion\Entity\CohesionSettingsInterface;
  *   links = {
  *     "in-use" = "/admin/cohesion/cohesion_font_library/{cohesion_font_library}/in_use",
  *     "collection" = "/admin/cohesion/cohesion_website_settings"
+ *   },
+ *   config_export = {
+ *     "id",
+ *     "label",
+ *     "json_values",
+ *     "json_mapper",
+ *     "last_entity_update",
+ *     "locked",
+ *     "modified",
+ *     "selectable",
+ *     "source"
  *   }
  * )
  */
@@ -48,13 +56,13 @@ class FontLibrary extends WebsiteSettingsEntityBase implements CohesionSettingsI
   const ASSET_GROUP_ID = 'website_settings';
 
   /**
-   * Return all the icons combined for the form[]
+   * Return all the icons combined for the form[].
    *
-   * @return array|\stdClass|string
+   * @return array|object|string
    */
   public function getResourceObject() {
-    /** @var WebsiteSettingsApi $send_to_api */
-    $send_to_api = $this->apiProcessorManager()->createInstance('website_settings_api');
+    /** @var \Drupal\cohesion_website_settings\Plugin\Api\WebsiteSettingsApi $send_to_api */
+    $send_to_api = $this->getApiPluginInstance();
 
     return $send_to_api->getFontGroup();
   }
@@ -63,8 +71,8 @@ class FontLibrary extends WebsiteSettingsEntityBase implements CohesionSettingsI
    * {@inheritdoc}
    */
   public function process() {
-    /** @var WebsiteSettingsApi $send_to_api */
-    $send_to_api = $this->apiProcessorManager()->createInstance('website_settings_api');
+    /** @var \Drupal\cohesion_website_settings\Plugin\Api\WebsiteSettingsApi $send_to_api */
+    $send_to_api = $this->getApiPluginInstance();
     $send_to_api->setEntity($this);
     $send_to_api->send();
     $send_to_api->getData();
@@ -88,12 +96,16 @@ class FontLibrary extends WebsiteSettingsEntityBase implements CohesionSettingsI
         $original_json_values = $original->getDecodedJsonValues();
         $json_values = $this->getDecodedJsonValues();
 
-        // Clear the previous font files if the new font library files are different or as no files,
+        // Set the entity label from the name inside the JSON.
+        $this->setlabel($json_values['name']);
+
+        // Clear the previous font files if the new font library files are different or as no files,.
         if (isset($json_values['fontFiles'])) {
           if ((isset($original_json_values['fontFiles']) && $json_values['fontFiles'] != $original_json_values['fontFiles'])) {
-            $this->clearFontFiles($original_json_values);
+            $this->clearFontFiles($original_json_values, $json_values);
           }
-        } else {
+        }
+        else {
           $this->clearFontFiles($original_json_values);
         }
       }
@@ -110,8 +122,16 @@ class FontLibrary extends WebsiteSettingsEntityBase implements CohesionSettingsI
     $this->process();
 
     // Invalidate settings endpoint shared cache entries.
-    $tags = ('font_libraries' == $this->id()) ? [$this->id(), 'font_stack',] : [$this->id()];
+    $tags = ('font_libraries' == $this->id()) ? [$this->id(), 'font_stack'] : [$this->id()];
     Cache::invalidateTags($tags);
+  }
+
+  /**
+   * Set a description.
+   */
+  public function setLabel($label) {
+    $this->label = $label;
+    return $this;
   }
 
   /**
@@ -125,24 +145,43 @@ class FontLibrary extends WebsiteSettingsEntityBase implements CohesionSettingsI
    * {@inheritdoc}
    */
   public function getInUseMessage() {
-    return ['message' => ['#markup' => t('This font library has been tracked as in use in the places listed below. You should not delete it until you have removed its use.'),],];
+    return ['message' => ['#markup' => t('This font library has been tracked as in use in the places listed below. You should not delete it until you have removed its use.')]];
   }
 
+  /**
+   *
+   */
   public function clearData() {
     $json_value = $this->getDecodedJsonValues();
     $this->clearFontFiles($json_value);
   }
 
   /**
-   * Clear font files for a given json values of a font library
+   * Clear font files for a given json values of a font library.
    *
    * @param $json_value
+   * @param $new_json_values
    */
-  private function clearFontFiles($json_value) {
+  private function clearFontFiles($json_value, $new_json_values = NULL) {
+    /** @var \Drupal\Core\File\FileSystemInterface $file_system */
+    $file_system = \Drupal::service('file_system');
     if (isset($json_value['fontFiles']) && is_array($json_value['fontFiles'])) {
       foreach ($json_value['fontFiles'] as $fontFile) {
         if (is_array($fontFile) && isset($fontFile['uri']) && file_exists($fontFile['uri'])) {
-          \Drupal::service('cohesion.local_files_manager')->deleteFileByURI($fontFile['uri']);
+
+          $should_delete = TRUE;
+          // Only delete if file have not the same path
+          if($new_json_values && isset($new_json_values['fontFiles']) && is_array($new_json_values['fontFiles'])) {
+            foreach ($new_json_values['fontFiles'] as $newFontFile) {
+              if($file_system->realpath($newFontFile['uri']) == $file_system->realpath($fontFile['uri'])) {
+                $should_delete = FALSE;
+              }
+            }
+          }
+
+          if($should_delete) {
+            \Drupal::service('cohesion.local_files_manager')->deleteFileByURI($fontFile['uri']);
+          }
         }
       }
     }
@@ -164,4 +203,5 @@ class FontLibrary extends WebsiteSettingsEntityBase implements CohesionSettingsI
     $this->modified = TRUE;
     $this->status = TRUE;
   }
+
 }
